added gui customization for Navbar Width added gui customization Scrollbar Size added gui customization Notification Duration added small ModeButton on player for about and so added changing user icon if localhost admin or anonymous access centralize ManageNavbar split Accounts tabs
