+++ react-client/src/components/Logs/Logs.tsx
@@ -30,8 +30,6 @@ import { ServerEventInterface } from '../../services/server-event-service'
 import { SessionInterface } from '../../services/session-service'
 import { allowHtml, defaultTooltipSettings, logsApiUrl } from '../../utils'
 import { showError } from '../../utils/notifications'
-import ManageNavbar from '../ManageNavbar/ManageNavbar'
-import { NavbarItems } from '../../services/navbar-items'
 
 const Logs = ({ i18n, main, session, sse }: { i18n: I18nInterface, main: MainInterface, session: SessionInterface, sse: ServerEventInterface }) => {
   const canModify = havePermission(session, Permissions.settings_modify)
@@ -96,7 +94,7 @@ const Logs = ({ i18n, main, session, sse }: { i18n: I18nInterface, main: MainInt
     session.setDocumentTitle('Logs')
     session.useSseAs('Logs')
     session.stopPlayerSse()
-    main.setNavbarValue(<ManageNavbar i18n={i18n} session={session} selectedKey={NavbarItems.Tools} />)
+    main.setNavbarItem(i18n, session, Logs.name)
   }, [])
 
   useEffect(() => {
