+++ react-client/src/components/Settings/Settings.tsx
@@ -22,7 +22,6 @@ import _ from 'lodash'
 import { useEffect, useState } from 'react'
 import { IconCheck, IconExclamationMark } from '@tabler/icons-react'
 
-import ManageNavbar from '../ManageNavbar/ManageNavbar'
 import { havePermission, Permissions } from '../../services/accounts-service'
 import { I18nInterface } from '../../services/i18n-service'
 import { MainInterface } from '../../services/main-service'
@@ -35,7 +34,6 @@ import NavigationSettings from './NavigationSettings'
 import RenderersSettings from './RenderersSettings'
 import TranscodingSettings from './TranscodingSettings'
 import { showError, showLoading, updateError, updateInfo, updateSuccess } from '../../utils/notifications'
-import { NavbarItems } from '../../services/navbar-items'
 
 export default function Settings({ i18n, main, session, sse }: { i18n: I18nInterface, main: MainInterface, session: SessionInterface, sse: ServerEventInterface }) {
   const [advancedSettings] = useLocalStorage<boolean>({
@@ -75,6 +73,7 @@ export default function Settings({ i18n, main, session, sse }: { i18n: I18nInter
     session.setDocumentTitle('Server Settings')
     session.useSseAs('Settings')
     session.stopPlayerSse()
+    main.setNavbarItem(i18n, session, Settings.name)
   }, [])
 
   useEffect(() => {
@@ -115,10 +114,6 @@ export default function Settings({ i18n, main, session, sse }: { i18n: I18nInter
         })
     }
   }, [canView, formSetValues])
-
-  useEffect(() => {
-    main.setNavbarValue(<ManageNavbar i18n={i18n} session={session} selectedKey={NavbarItems.ServerSettings} />)
-  }, [i18n.get, main.setNavbarValue])
 
   const handleSubmit = async (values: typeof form.values) => {
     setLoading(true)
